pyproject.toml
[project]
name = "iacc-prgav1301"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.1.1",
]
